mature-evaluation-system-backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
Pillow==10.1.0
reportlab==4.0.7
django-cors-headers==4.3.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
openpyxl==3.1.2
drf-yasg==1.21.7
openai==1.3.7
pypinyin==0.50.0
httpx==0.27.0
PyPDF2
python-docx
xhtml2pdf
matplotlib
django-storages[boto3]==1.14.2
oss2==2.18.4